Turn off decryption
Former-commit-id: 09bcda270547dd22bc07074911346094e19b9461

+++ connect.go
@@ -87,7 +87,9 @@ func runClient(connectionType string, codePhrase string) {
 			return
 		}
 		ioutil.WriteFile(fileName, decrypted, 0644)
-		os.Remove(fileName + ".encrypted")
+		if !debugFlag {
+			os.Remove(fileName + ".encrypted")
+		}
 		log.Debugf("\n\n\ndownloaded hash: [%s]", HashBytes(decrypted))
 		log.Debugf("\n\n\nrelayed hash: [%s]", fileHash)
